Refactor 2b763e3 for readability
I noticed myself performing repeated regressive saccades due to the
parameter renaming in the if-statement which this commit removes.

The arguments array is copied to remain strictness-neutral, in keeping
with the existing implementation.

The choice of argument assignment on a type check makes sense as the
`Dom` methods are expected to be frequently called with only two or three
arguments. Note that `DomListener` effectively shims the proposed
[DOM Level 3 Events] specification by explicitly setting `useCapture`
to `false` when undefined for compatibility with [DOM 2] browsers.

[DOM Level 3 Events]: https://dvcs.w3.org/hg/dom3events/raw-file/tip/html/DOM3-Events.html
[DOM 2]: http://www.w3.org/TR/DOM-Level-2-Events/events.html