options.sourceMapsInProduction => options.disableScriptMap
Maps should be on by default, so that people can debug in production.
The option should be called scriptMap for consistency with other
variable names. (CSS can also have source maps, so script map is more
specific.) I don't think we should hardcode the logic for production
vs. not production. This is easy enough for someone using this function
to do, and they might want to have different rules for different
environments.