Switch to immediate model listeners
This fixes a bug in binding updates, triggered when a model listener causes a change to the path being listened on.

Preliminary app testing all seems fine, but to be safe, I'm putting this behind a flag. To enable the new behavior, do a `model.set('$derbyFlags.immediateModelListeners', true)` before calling `app.createPage()`.